# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LunamHub Family Command Centre API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rewards:
    get:
      operationId: listRewards
      tags: [rewards]
      summary: List reward items
      responses:
        "200":
          description: Rewards list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reward"
    post:
      operationId: createReward
      tags: [rewards]
      summary: Create a reward
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RewardInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reward"
  /rewards/{id}:
    patch:
      operationId: updateReward
      tags: [rewards]
      summary: Update a reward
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RewardUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reward"
    delete:
      operationId: deleteReward
      tags: [rewards]
      summary: Delete a reward
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /routines:
    get:
      operationId: listRoutines
      tags: [routines]
      summary: List routines
      responses:
        "200":
          description: Routines list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Routine"
    post:
      operationId: createRoutine
      tags: [routines]
      summary: Create a routine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoutineInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Routine"
  /routines/{id}:
    get:
      operationId: getRoutine
      tags: [routines]
      summary: Get a routine with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Routine with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutineWithItems"
        "404":
          description: Not found
    patch:
      operationId: updateRoutine
      tags: [routines]
      summary: Update a routine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoutineUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Routine"
    delete:
      operationId: deleteRoutine
      tags: [routines]
      summary: Delete a routine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /routines/{id}/items:
    post:
      operationId: createRoutineItem
      tags: [routines]
      summary: Add an item to a routine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoutineItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutineItem"
  /routines/{id}/items/{itemId}:
    patch:
      operationId: updateRoutineItem
      tags: [routines]
      summary: Update a routine item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoutineItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutineItem"
    delete:
      operationId: deleteRoutineItem
      tags: [routines]
      summary: Delete a routine item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /routines/{id}/complete:
    post:
      operationId: completeRoutineItem
      tags: [routines]
      summary: Mark a routine item as complete for today
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoutineCompletionInput"
      responses:
        "200":
          description: Completion recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutineCompletion"

  # ── Dashboard ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    FamilyMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        emoji:
          type: string
        color:
          type: string
        role:
          type: string
          enum: [parent, child]
        pointsBalance:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - emoji
        - color
        - role
        - pointsBalance
        - createdAt
...
    Reward:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        pointsCost:
          type: integer
        active:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - pointsCost
        - active
        - createdAt
    RewardInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        pointsCost:
          type: integer
        active:
          type: boolean
      required:
        - title
        - pointsCost
    RewardUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        pointsCost:
          type: integer
        active:
          type: boolean
...
    Routine:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        assignedTo:
          type: ["integer", "null"]
        assignedMember:
          $ref: "#/components/schemas/FamilyMember"
        timeOfDay:
          type: string
          enum: [morning, afternoon, evening, bedtime]
        createdAt:
          type: string
      required:
        - id
        - name
        - timeOfDay
        - createdAt
    RoutineInput:
      type: object
      properties:
        name:
          type: string
        assignedTo:
          type: integer
        timeOfDay:
          type: string
          enum: [morning, afternoon, evening, bedtime]
      required:
        - name
        - timeOfDay
    RoutineUpdate:
      type: object
      properties:
        name:
          type: string
        assignedTo:
          type: ["integer", "null"]
        timeOfDay:
          type: string
          enum: [morning, afternoon, evening, bedtime]
    RoutineItem:
      type: object
      properties:
        id:
          type: integer
        routineId:
          type: integer
        text:
          type: string
        order:
          type: integer
      required:
        - id
        - routineId
        - text
        - order
    RoutineItemInput:
      type: object
      properties:
        text:
          type: string
        order:
          type: integer
      required:
        - text
    RoutineItemUpdate:
      type: object
      properties:
        text:
          type: string
        order:
          type: integer
    RoutineWithItems:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        assignedTo:
          type: ["integer", "null"]
        timeOfDay:
          type: string
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/RoutineItem"
        completionsToday:
          type: array
          items:
            type: integer
      required:
        - id
        - name
        - timeOfDay
        - createdAt
        - items
        - completionsToday
    RoutineCompletion:
      type: object
      properties:
        id:
          type: integer
        routineId:
          type: integer
        routineItemId:
          type: integer
        completedAt:
          type: string
      required:
        - id
        - routineId
        - routineItemId
        - completedAt
    RoutineCompletionInput:
      type: object
      properties:
        routineItemId:
          type: integer
      required:
        - routineItemId

    # ── Dashboard ─────────────────────────────────────────────────────────────
...