# lib/api-zod/src/generated/types/routineUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LunamHub Family Command Centre API
 * OpenAPI spec version: 0.1.0
 */
import type { RoutineUpdateTimeOfDay } from './routineUpdateTimeOfDay';

export interface RoutineUpdate {
  name?: string;
  /** @nullable */
  assignedTo?: number | null;
  timeOfDay?: RoutineUpdateTimeOfDay;
}